+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModInnerPlatform.kt
@@ -11,7 +11,6 @@ import site.siredvin.peripheralworks.networking.ServerNetworkContext
 import site.siredvin.tweakium.modules.platform.api.InnerComputerBasePlatform
 
 interface ModInnerPlatform : InnerComputerBasePlatform {
-    val commonEnergy: EnergyUnit
     val modList: List<String>
     fun getModInformation(mod: String): Map<String, Any>?