+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/powah/PowahEnergyStorageWrapper.kt
@@ -2,6 +2,7 @@ package site.siredvin.peripheralworks.integrations.powah
 
 import owmii.powah.lib.block.AbstractEnergyBlock
 import owmii.powah.lib.block.AbstractEnergyStorage
+import site.siredvin.broccolium.modules.platform.PlatformToolkit
 import site.siredvin.broccolium.modules.storage.energy.AgnosticEnergyStack
 import site.siredvin.broccolium.modules.storage.energy.EnergyUnit
 import site.siredvin.broccolium.modules.storage.energy.api.AgnosticEnergyStorage
@@ -14,7 +15,7 @@ class PowahEnergyStorageWrapper(private val storage: AbstractEnergyStorage<*, *>
     override val canExtract: Boolean
         get() = storage.canExtractEnergy(null)
     override val energy: AgnosticEnergyStack
-        get() = AgnosticEnergyStack(ModPlatform.commonEnergy, storage.energy.stored)
+        get() = AgnosticEnergyStack(PlatformToolkit.get().commonEnergy, storage.energy.stored)
 
     override val receiveRateLimit: Long
         get() = (storage.block as? AbstractEnergyBlock)?.config?.getTransfer(storage.variant) ?: Long.MAX_VALUE
@@ -29,19 +30,19 @@ class PowahEnergyStorageWrapper(private val storage: AbstractEnergyStorage<*, *>
     override val canReceive: Boolean
         get() = storage.canReceiveEnergy(null)
     override val unit: EnergyUnit
-        get() = ModPlatform.commonEnergy
+        get() = PlatformToolkit.get().commonEnergy
 
     override fun storeEnergy(stack: AgnosticEnergyStack): AgnosticEnergyStack {
-        if (!stack.`is`(ModPlatform.commonEnergy)) return stack
+        if (!stack.`is`(PlatformToolkit.get().commonEnergy)) return stack
         if (!storage.canReceiveEnergy(null)) return stack
         val receivedEnergy = storage.receiveEnergy(stack.amount, false, null)
         return stack.copyWithCount(stack.amount - receivedEnergy)
     }
 
     override fun takeEnergy(predicate: Predicate<AgnosticEnergyStack>, limit: Long): AgnosticEnergyStack {
-        if (!predicate.test(energy)) return AgnosticEnergyStack(ModPlatform.commonEnergy, 0)
+        if (!predicate.test(energy)) return AgnosticEnergyStack(PlatformToolkit.get().commonEnergy, 0)
         val extractedEnergy = storage.extractEnergy(limit, false, null)
-        if (extractedEnergy == 0L) return AgnosticEnergyStack(ModPlatform.commonEnergy, 0)
-        return AgnosticEnergyStack(ModPlatform.commonEnergy, extractedEnergy)
+        if (extractedEnergy == 0L) return AgnosticEnergyStack(PlatformToolkit.get().commonEnergy, 0)
+        return AgnosticEnergyStack(PlatformToolkit.get().commonEnergy, extractedEnergy)
     }
 }